Add the trading shop back into Shoppy

+++ src/main/java/com/epherical/shoppy/network/payloads/PriceSubmissionPayload.java
@@ -11,19 +11,18 @@ import net.minecraft.resources.ResourceLocation;
  * Sent when the owner finishes editing the offer-price text boxes.
  *
  * @param offerIndex index 0‒2 of the offer that is being edited
- * @param price      currency items required (0 = discard server-side)
+ * @param price      currency amount required (0 = discard server-side)
  * @param received   items the buyer will get
  */
-public record PriceSubmissionPayload(int offerIndex, int price, int received) implements CustomPacketPayload {
+public record PriceSubmissionPayload(int offerIndex, double price, int received) implements CustomPacketPayload {
 
     public static final ResourceLocation ID = ResourceLocation.fromNamespaceAndPath(Shoppy.MODID, "price_submission");
 
     public static final Type<PriceSubmissionPayload> TYPE = new Type<>(ID);
 
-    /* ───────────────────────────── codec ───────────────────────────── */
     public static final StreamCodec<RegistryFriendlyByteBuf, PriceSubmissionPayload> STREAM_CODEC = StreamCodec.composite(
             ByteBufCodecs.INT, PriceSubmissionPayload::offerIndex,
-            ByteBufCodecs.INT, PriceSubmissionPayload::price,
+            ByteBufCodecs.DOUBLE, PriceSubmissionPayload::price,
             ByteBufCodecs.INT, PriceSubmissionPayload::received,
             PriceSubmissionPayload::new);
